requirements.txt
absl-py==2.1.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
AutoDockTools_py3==1.5.7.post1+12.ga62a5d9
Babel==2.15.0
beautifulsoup4==4.12.3
biopython==1.85
bleach==6.1.0
brokenaxes==0.6.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
comm==0.2.2
conda-pack==0.7.1
contextlib2==21.6.0
contourpy==1.2.1
cupy-cuda12x==13.4.1
cutensor-cu12==2.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
etils==1.7.0
evox==0.9.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fastrlock==0.8.3
filelock==3.14.0
flatbuffers==25.2.10
flax==0.8.3
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.5.0
gast==0.6.0
gemmi==0.7.3
google-pasta==0.2.0
grpcio==1.74.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.11.0
kiwisolver==1.4.5
Levenshtein==0.25.1
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meeko==0.6.1
mistune==3.0.2
ml-collections==0.1.1
ml_dtypes==0.5.3
MolVS==0.1.1
msgpack==1.0.8
namex==0.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
openbabel-wheel==3.1.1.21
opt-einsum==3.3.0
optax==0.2.2
optree==0.17.0
orbax-checkpoint==0.5.11
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdb-tools==2.5.0
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pyscf==2.9.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.7
rdkit==2023.9.6
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
selfies==2.1.1
Send2Trash==1.8.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.59
termcolor==3.1.0
terminado==0.18.1
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.17.2
zipp==3.18.1
